+++ docs/sections/api/datatypes/properties/dim3/elastic_isotropic.rst
@@ -1,0 +1,9 @@
+.. _point_properties_dim3_elastic_isotropic:
+
+Point Properties (3D Elastic Isotropic)
+=======================================
+
+.. doxygengroup:: specfem_point_properties_dim3_elastic_isotropic
+    :members:
+    :private-members:
+    :content-only:
