+++ tests/unit-tests/CMakeLists.txt
@@ -115,43 +115,36 @@ target_link_libraries(
   -lpthread -lm
 )
 
-add_executable(
-  point_coordinates_tests
-  point/coordinates_tests.cpp
-)
-target_link_libraries(
-  point_coordinates_tests
-  point
-  gtest_main
-  gmock_main
-)
 
 add_executable(
-  point_index_tests
+  point_tests
   point/index_tests.cpp
+  point/coordinates_tests.cpp
+  # Kernels
+  # Dim 2
+  point/kernels/dim2/acoustic_isotropic.cpp
+  point/kernels/dim2/elastic_isotropic.cpp
+  point/kernels/dim2/elastic_anisotropic.cpp
+  point/kernels/dim2/poroelastic_isotropic.cpp
+  # Properties
+  # Dim 2
+  point/properties/dim2/elastic_isotropic.cpp
+  point/properties/dim2/elastic_anisotropic.cpp
+  point/properties/dim2/acoustic_isotropic.cpp
+  point/properties/dim2/electromagnetic_isotropic.cpp
+  point/properties/dim2/poroelastic_isotropic.cpp
+  # Dim 3
+  point/properties/dim3/elastic_isotropic.cpp
+
 )
+
 target_link_libraries(
-  point_index_tests
+  point_tests
   point
   gtest_main
   gmock_main
 )
 
-add_executable(
-  point_properties_tests
-  point/properties_tests.cpp
-)
-target_link_libraries(
-  point_properties_tests
-  point
-  gtest_main
-  gmock_main
-)
-
-# add_executable(
-#   mesh_tests
-#   mesh/mesh_tests.cpp
-# )
 
 add_executable(
   mesh_tests
@@ -476,9 +469,7 @@ if (NOT MPI_PARALLEL)
   # gtest_discover_tests(compute_elastic_tests WORKING_DIRECTORY ${TEST_DIR})
   # # gtest_discover_tests(compute_acoustic_tests WORKING_DIRECTORY ${TEST_DIR})
   # gtest_discover_tests(compute_coupled_interfaces_tests WORKING_DIRECTORY ${TEST_DIR})
-  gtest_discover_tests(point_coordinates_tests WORKING_DIRECTORY ${TEST_DIR})
-  gtest_discover_tests(point_index_tests WORKING_DIRECTORY ${TEST_DIR})
-  gtest_discover_tests(point_properties_tests WORKING_DIRECTORY ${TEST_DIR})
+  gtest_discover_tests(point_tests WORKING_DIRECTORY ${TEST_DIR})
   gtest_discover_tests(compute_tests WORKING_DIRECTORY ${TEST_DIR})
   gtest_discover_tests(assembly_tests WORKING_DIRECTORY ${TEST_DIR})
   gtest_discover_tests(policies WORKING_DIRECTORY ${TEST_DIR})
